+++ CMakeLists.txt
@@ -1,12 +1,15 @@
 cmake_minimum_required(VERSION 3.15)
 project(Piano)
-add_executable(piano src/main.cpp)
-find_path(RAYGUI_INCLUDE_DIRS "include/raygui/raygui.h")
-target_include_directories(piano PRIVATE ${RAYGUI_INCLUDE_DIRS})
+add_executable(piano src/main.cpp src/piano.cpp src/chord.cpp)
+
+include_directories(${CMAKE_SOURCE_DIR}/include)
+
 find_package(RtMidi REQUIRED)
 target_link_libraries(piano RtMidi::rtmidi)
+
 find_package(raylib REQUIRED)
 target_link_libraries(piano raylib)
+
 find_library(OpenGL_LIBRARY OpenGL)
 find_library(COCOA_LIBRARY Cocoa)
 find_library(IOKIT_LIBRARY IOKit)
